End of section 18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_modals/message';
 import { Pagination } from '../_modals/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,7 +17,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -31,11 +32,29 @@ export class MessagesComponent implements OnInit {
     })
   }
 
-  deleteMassage(ID: number) {
-    this.messageService.deleteMessage(ID).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id = ID), 1);
+  deleteMessage(ID: number) {
+    // this.messageService.deleteMessage(id).subscribe(() => {
+    //   this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+    this.confirmService.confirm('Confirm delete message', 'This cannot be undone').subscribe(result => {
+      if (result) {
+        this.messageService.deleteMessage(ID).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === ID), 1);
+        })
+      }
     })
+
   }
+
+  // deleteMassage(ID: number) {
+  //   this.confirmService.confirm('Confirm delete message', 'This cannont be undone').subscribe(result => {
+  //     if(result) {
+  //       this.messageService.deleteMessage(ID).subscribe(() => {
+  //         this.messages.splice(this.messages.findIndex(m => m.id = ID), 1);
+  //       })
+  //     }
+  //   })
+    
+  // }
   pageChanged(event: any) {
     this.pageNumber = event.page;
     this.loadMessages();
